// back/src/types/evolution.d.ts
export interface Evolution {
    ID?                     : number
    GENDER                  : number
    HELD_ITEM               : string
    ITEM                    : string
    KNOW_MOVE               : string
    KNOW_MOVE_TYPE          : string
    LOCATION                : string
    MIN_AFFECTION           : number
    MIN_BEAUTY              : number
    MIN_HAPPINESS           : number
    MIN_LEVEL               : number
    NEEDS_OVERWORLDS_RAIN   : number
    PARTY_SPECIES           : string
    PARTY_TYPE              : string
    RELATIVE_PHYSICAL_STATS : number
    TIME_OF_DAY             : string
    TRADE_SPECIES           : string
    TRIGGER                 : string
    POKEMON_ID              : number
    CHAIN_ID                : number
    LEVEL                   : number
}